Add new component for Sinclair AC IR remote
This is a basic implementation of the Wireless remote controller Y512SB
for Sinclair AC unit. The component can receive and transmit, although
only the receive function has been tested and is working. More work and
testing are required for the transmit function.

The component can set temperature, working mode and fan speed. The
swing, sleep, timer and purify functions are not implemented, yet.